vsock: move testing utilities in separate module
